Add a gate job for deployments
Prior to this change, we did not build the image for a deployment until
all tests had passed. However, this meant waiting for all tests to pass
before starting to build the image.

This change builds the image immediately and then waits until all of the
tests and linters have passed before deploying.

This will speed up the deployment process. It is generally safe to try
to build immediately because we only merge once all of the tests and
linters have passed and there's only one of me working on the project,
so difficulties arising from not rebasing will be rare.