Return a useful message when no information is found for an ISBN
Prior to this change, we returned an empty 404 response when no
information could be found for the ISBN. However, iOS shortcuts does not
diplsya the response status, so it was impossible to know in the
shortcut whether the request had failed or why.

This change shows the exception message if the event could not be
recorded.